sampleone/sample/requirements.txt
flask
scikit-learn
pandas
numpy
joblib
matplotlib
mysql-connector-python
